+++ tests/unit/test_helper_dish_ln_device.py
@@ -59,6 +59,18 @@ def test_dish_ln_commands_scan(tango_context):
     assert message[0] == ""
 
 
+def test_scan_command_without_argin_failed_result(tango_context):
+    dev_factory = DevFactory()
+    dish_device = dev_factory.get_device(DISH_LN_DEVICE)
+    dish_device.SetDefective(json.dumps(FAILED_RESULT_DEFECT))
+    result, message = dish_device.command_inout("Scan", "")
+    assert result[0] == ResultCode.FAILED
+    assert (
+        message[0] == "Device is defective, cannot process command completely."
+    )
+    dish_device.SetDefective(json.dumps({"enabled": False}))
+
+
 @pytest.mark.parametrize("command_to_check", COMMANDS_WITHOUT_INPUT)
 def test_command_without_argin_failed_result(tango_context, command_to_check):
     dev_factory = DevFactory()